requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
PyYAML>=6.0.0
scikit-learn>=1.3.0
yt-dlp>=2024.0.0